api/requirements.txt
aioredis==2.0.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
click==8.1.7
dnspython==2.4.2
exceptiongroup==1.1.3
fastapi==0.101.1
h11==0.14.0
idna==3.4
motor==3.2.0
paho-mqtt==1.6.1
pydantic==2.2.1
pydantic_core==2.6.1
PyJWT==2.8.0
pymongo==4.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
redis==5.0.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
